+++ molo/surveys/blocks.py
@@ -1,6 +1,7 @@
 from django import forms
 from django.contrib.staticfiles.templatetags.staticfiles import static
 from django.core.exceptions import ValidationError
+from django.utils.translation import ugettext_lazy as _
 
 from wagtail.wagtailadmin.edit_handlers import StreamFieldPanel
 from wagtail.wagtailcore import blocks
@@ -14,11 +15,14 @@ class SkipState:
     SURVEY = 'survey'
 
 
+VALID_SKIP_SELECTORS = ['radio', 'checkbox', 'dropdown']
+
+
 class SkipLogicField(StreamField):
     def __init__(self, *args, **kwargs):
         args = [SkipLogicStreamBlock([('skip_logic', SkipLogicBlock())])]
         kwargs.update({
-            'verbose_name': 'Answer options',
+            'verbose_name': _('Answer options'),
             'blank': True,
         })
         super(SkipLogicField, self).__init__(*args, **kwargs)
@@ -75,10 +79,10 @@ class SkipLogicBlock(blocks.StructBlock):
     choice = blocks.CharBlock()
     skip_logic = blocks.ChoiceBlock(
         choices=[
-            (SkipState.NEXT, 'Next default question'),
-            (SkipState.END, 'End of survey'),
-            (SkipState.QUESTION, 'Another question'),
-            (SkipState.SURVEY, 'Another survey'),
+            (SkipState.NEXT, _('Next default question')),
+            (SkipState.END, _('End of survey')),
+            (SkipState.QUESTION, _('Another question')),
+            (SkipState.SURVEY, _('Another survey')),
         ],
         default=SkipState.NEXT,
         required=True,
@@ -89,8 +93,10 @@ class SkipLogicBlock(blocks.StructBlock):
     )
     question = QuestionSelectBlock(
         required=False,
-        help_text=('Please save the survey as a draft to populate or update '
-                   'the list of questions.'),
+        help_text=_(
+            ('Please save the survey as a draft to populate or update '
+             'the list of questions.')
+        ),
     )
 
     @property
@@ -98,7 +104,7 @@ class SkipLogicBlock(blocks.StructBlock):
         return forms.Media(js=[static('js/blocks/skiplogic.js')])
 
     def js_initializer(self):
-        opts = {'validSkipSelectors': ['radio', 'checkbox', 'dropdown']}
+        opts = {'validSkipSelectors': VALID_SKIP_SELECTORS}
         return "SkipLogic(%s)" % blocks.utils.js_dict(opts)
 
     def clean(self, value):
@@ -108,7 +114,7 @@ class SkipLogicBlock(blocks.StructBlock):
             if not cleaned_data['survey']:
                 raise ValidationError(
                     'A Survey must be selected to progress to.',
-                    params={'survey': ['Please select a survey.']}
+                    params={'survey': [_('Please select a survey.')]}
                 )
             cleaned_data['question'] = None
 
@@ -116,7 +122,7 @@ class SkipLogicBlock(blocks.StructBlock):
             if not cleaned_data['question']:
                 raise ValidationError(
                     'A Question must be selected to progress to.',
-                    params={'question': ['Please select a question.']}
+                    params={'question': [_('Please select a question.')]}
                 )
             cleaned_data['survey'] = None