+++ CHANGES.rst
@@ -5,6 +5,7 @@ Upcoming
 --------
 
 - Added Article Tag Rule to allow segmenting on article visits
+- Added ability to skip questions and surveys based on user's response
 
 5.6.2
 -----